chore: upgrade dependencies to latest
