+++ package.json
@@ -1,55 +1,66 @@
 {
-	"name": "astro-selfie",
-	"version": "1.0.2",
-	"description": "Astro integration to generate page screenshots to show as Open Graph images",
-	"license": "ISC",
-	"repository": "vadimdemedes/astro-selfie",
-	"author": {
-		"name": "Vadim Demedes",
-		"email": "vadimdemedes@hey.com",
-		"url": "https://github.com/vadimdemedes"
-	},
-	"type": "module",
-	"exports": {
-		"types": "./dist/index.d.ts",
-		"default": "./dist/index.js"
-	},
-	"engines": {
-		"node": ">=14.16"
-	},
-	"scripts": {
-		"build": "tsc",
-		"dev": "tsc --watch",
-		"test": "prettier --check source && xo",
-		"prepare": "rm -rf dist && tsc"
-	},
-	"files": [
-		"dist"
-	],
-	"keywords": [
-		"astro-integration"
-	],
-	"dependencies": {
-		"get-port": "^7.0.0",
-		"micro": "^10.0.1",
-		"playwright": "^1.34.3",
-		"serve-handler": "^6.1.5"
-	},
-	"devDependencies": {
-		"@sindresorhus/tsconfig": "^3.0.1",
-		"@types/react": "^18.2.9",
-		"@types/serve-handler": "^6.1.1",
-		"@vdemedes/prettier-config": "^2.0.1",
-		"astro": "^2.6.1",
-		"prettier": "^2.8.8",
-		"typescript": "^5.1.3",
-		"xo": "^0.54.2"
-	},
-	"prettier": "@vdemedes/prettier-config",
-	"xo": {
-		"prettier": true,
-		"rules": {
-			"no-await-in-loop": "off"
-		}
-	}
+  "name": "@celestio/astro-selfie",
+  "version": "1.0.3",
+  "description": "Astro integration to generate page screenshots to show as Open Graph images",
+  "license": "ISC",
+  "author": {
+    "name": "Yann V.",
+    "email": "yann@celest.io",
+    "url": "https://github.com/celest-io"
+  },
+  "homepage": "https://github.com/celest-io/astro-selfi#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/celest-io/astro-selfi.git"
+  },
+  "type": "module",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    "types": "./dist/index.d.ts",
+    "default": "./dist/index.js"
+  },
+  "engines": {
+    "node": ">=22",
+    "pnpm": ">=10"
+  },
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsc --watch",
+    "test": "xo",
+    "clean": "rm -rf dist",
+    "prepare": "rm -rf dist && tsc"
+  },
+  "files": [
+    "dist"
+  ],
+  "keywords": [
+    "astro-integration",
+    "astro"
+  ],
+  "dependencies": {
+    "@playwright/browser-chromium": "^1.54.2",
+    "get-port": "^7.1.0",
+    "micro": "^10.0.1",
+    "playwright": "^1.54.2",
+    "serve-handler": "^6.1.6"
+  },
+  "devDependencies": {
+    "@types/react": "^19.1.9",
+    "@types/serve-handler": "^6.1.4",
+    "astro": "^5.12.0",
+    "prettier": "^3.6.2",
+    "typescript": "^5.9.2",
+    "xo": "^1.2.1"
+  },
+  "publishConfig": {
+    "provenance": true
+  },
+  "xo": {
+    "space": true,
+    "prettier": true,
+    "rules": {
+      "no-await-in-loop": "off"
+    }
+  }
 }
